Extract Msf::DBManager::Import::Spiceworks
MSP-11124
